refac: Move formatUnitsToFixed
Move formatUnitsToFixed from app/page/info/tokens.tsx to
packages/network/util/index.ts.
